Prevent error when path is empty
When the homepage doesn't exists a redirect for "/" is fetched. When it
can't find that it fetched path empty string "". DynamoDB doesn't except
empty string in the query and it's useless. Because redirecting empty
string to something will never happen. So skip the API call when path is
an empty string.